tests: fixed the move logic
